Fix `karere::terminate()` without session

When `MegaChatApi::logout()` --> `karere::terminate()` is called without a session, the `wipeDB(null)` fails the assertion.

See merge request !44

+++ src/chatClient.cpp
@@ -758,7 +758,7 @@ promise::Promise<void> Client::terminate(bool deleteDb)
     return disconnect()
     .then([this, deleteDb]()
     {
-        if (deleteDb)
+        if (deleteDb && !mSid.empty())
         {
             wipeDb(mSid);
         }
